refactor: Room Instance 생성 코드 리펙토링 https://github.com/Team-PATCH/Harmony-Android/pull/148#discussion_r2068558263
