+++ core/database/src/main/java/com/harmony/core/database/HarmonyDatabase.kt
@@ -1,10 +1,7 @@
 package com.harmony.core.database
 
-import android.content.Context
 import androidx.room.Database
-import androidx.room.Room
 import androidx.room.RoomDatabase
-import androidx.sqlite.db.SupportSQLiteDatabase
 import com.harmony.core.database.dao.GroupDao
 import com.harmony.core.database.dao.QuestionDao
 import com.harmony.core.database.dao.TodoDao
@@ -14,8 +11,6 @@ import com.harmony.core.database.model.QuestionCommentEntity
 import com.harmony.core.database.model.QuestionEntity
 import com.harmony.core.database.model.TodoEntity
 import com.harmony.core.database.model.UserEntity
-import com.harmony.core.database.model.preload.QuestionPreloadData
-import com.harmony.core.database.model.preload.TodoPreloadData
 
 @Database(
     entities = [
@@ -34,33 +29,6 @@ internal abstract class HarmonyDatabase : RoomDatabase() {
     abstract fun questionDao(): QuestionDao
 
     companion object {
-        private const val DB_NAME = "harmony.db"
-
-        @Volatile
-        private var instance: HarmonyDatabase? = null
-
-        fun getInstance(context: Context): HarmonyDatabase {
-            val roomDatabaseCallback: Callback = object : Callback() {
-                override fun onCreate(db: SupportSQLiteDatabase) {
-                    super.onCreate(db)
-                    listOf(
-                        QuestionPreloadData(),
-                        TodoPreloadData()
-                    )
-                        .forEach { it.insertPreloadData(db) }
-                }
-            }
-
-            return instance ?: synchronized(HarmonyDatabase::class) {
-                instance ?: Room.databaseBuilder(
-                    context.applicationContext,
-                    HarmonyDatabase::class.java,
-                    DB_NAME
-                )
-                    .addCallback(roomDatabaseCallback)
-                    .build()
-                    .also { instance = it }
-            }
-        }
+        internal const val DB_NAME = "harmony.db"
     }
 }
